requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

captum==0.7.0
imageio==2.33.0
matplotlib==3.8.2
mlflow==2.9.1
numpy==1.26.2
opencv-python==4.8.1.78
pandas==2.1.3
pypdf==3.17.2
pytest==7.4.3
PyYAML==6.0.1
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
shap==0.44.0
termcolor==2.4.0
tqdm==4.66.1
xgboost==2.0.2
scikit-learn==1.3.2
chardet==5.2.0
notebook
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
-f https://download.pytorch.org/whl/cu121
